rename extension candidate resolution
